+++ copy_with_extension/example/example.dart
@@ -1,39 +1,36 @@
-// ignore_for_file: unused_element
+// ignore_for_file: unused_element, unused_element_parameter
 
 import 'package:copy_with_extension/copy_with_extension.dart';
 
-/// Make sure the `part` is specified before running the builder.
-/// part 'example.g.dart'; /// It should not be commented.
+/// Make sure the `part` directive is specified before running the builder.
+/// part 'example.g.dart'; // Do not comment this line.
 
-/// Lets you use it like this: `SimpleObject(id: "test").copyWith(id: "new values", intValue: 10).copyWithNull(intValue: true)`.
+/// Allows you to use it like this: `SimpleObject(id: "test").copyWith(id: "new values", intValue: 10).copyWithNull(intValue: true)`.
 /// Or like this: `SimpleObject(id: "test").copyWith.id("new value")`.
 @CopyWith(copyWithNull: true)
 class SimpleObjectOldStyle {
+  const SimpleObjectOldStyle({required this.id, this.intValue});
+
   final String id;
   final int? intValue;
-
-  /// Make sure that constructor has named parameters (wrapped in curly braces)
-  const SimpleObjectOldStyle({required this.id, this.intValue});
 }
 
-/// Won't allow you to copy this object with a modified `id` field after object creation. It will always copy it from the original instance.
+/// Prevents modification of the `id` field after object creation. The value is always copied from the original instance.
 @CopyWith()
 class SimpleObjectImmutableField {
+  const SimpleObjectImmutableField({this.id, this.intValue});
+
   @CopyWithField(immutable: true)
   final String? id;
   final int? intValue;
-
-  /// Make sure that constructor has named parameters (wrapped in curly braces)
-  const SimpleObjectImmutableField({this.id, this.intValue});
 }
 
 /// Allows the use of a private constructor.
 @CopyWith(constructor: "_")
 class SimpleObjectPrivateConstructor {
+  const SimpleObjectPrivateConstructor._({this.id, this.intValue});
+
   @CopyWithField(immutable: true)
   final String? id;
   final int? intValue;
-
-  /// Make sure that constructor has named parameters (wrapped in curly braces)
-  const SimpleObjectPrivateConstructor._({this.id, this.intValue});
 }