+++ copy_with_extension_gen/lib/src/copy_with_generator.dart
@@ -1,115 +1,82 @@
-import 'package:analyzer/dart/element/element.dart' show ClassElement, Element;
+import 'package:analyzer/dart/element/element2.dart'
+    show ClassElement2, Element2;
 import 'package:build/build.dart' show BuildStep;
 import 'package:copy_with_extension/copy_with_extension.dart';
 import 'package:copy_with_extension_gen/src/field_info.dart';
 import 'package:copy_with_extension_gen/src/helpers.dart';
+import 'package:copy_with_extension_gen/src/settings.dart';
 import 'package:source_gen/source_gen.dart'
     show ConstantReader, GeneratorForAnnotation, InvalidGenerationSourceError;
 
 /// A `Generator` for `package:build_runner`
 class CopyWithGenerator extends GeneratorForAnnotation<CopyWith> {
+  CopyWithGenerator(this.settings) : super();
+
+  Settings settings;
+
   @override
   String generateForAnnotatedElement(
-    Element element,
+    Element2 element,
     ConstantReader annotation,
     BuildStep buildStep,
   ) {
-    if (element is! ClassElement) {
+    if (element is! ClassElement2) {
       throw InvalidGenerationSourceError(
         'Only classes can be annotated with "CopyWith". "$element" is not a ClassElement.',
         element: element,
       );
     }
 
-    final ClassElement classElement = element;
-    final classAnnotation = readClassAnnotation(annotation);
+    final ClassElement2 classElement = element;
+    final privacyPrefix = element.isPrivate ? "_" : "";
+    final classAnnotation = readClassAnnotation(settings, annotation);
 
+    final className = readElementNameOrThrow(classElement);
     final sortedFields =
         sortedConstructorFields(classElement, classAnnotation.constructor);
     final typeParametersAnnotation = typeParametersString(classElement, false);
     final typeParametersNames = typeParametersString(classElement, true);
-    final typeAnnotation = classElement.name + typeParametersNames;
+    final typeAnnotation = className + typeParametersNames;
+
+    for (final field in sortedFields) {
+      if (field.classFieldInfo != null &&
+          field.nullable == false &&
+          field.classFieldInfo?.nullable == true) {
+        throw InvalidGenerationSourceError(
+          'The constructor parameter "${field.name}" is not nullable, whereas the corresponding class field is nullable. This use case is not supported.',
+          element: element,
+        );
+      }
+    }
 
     return '''
     ${_copyWithProxyPart(
       classAnnotation.constructor,
-      classElement.name,
+      className,
       typeParametersAnnotation,
       typeParametersNames,
       sortedFields,
       classAnnotation.skipFields,
     )}
     
-    extension \$${classElement.name}CopyWith$typeParametersAnnotation on ${classElement.name}$typeParametersNames {
-      /// Returns a callable class that can be used as follows: `instanceOf${classElement.name}.copyWith(...)`${classAnnotation.skipFields ? "" : " or like so:`instanceOf${classElement.name}.copyWith.fieldName(...)`"}.
+    extension $privacyPrefix\$${className}CopyWith$typeParametersAnnotation on $className$typeParametersNames {
+      /// Returns a callable class that can be used as follows: `instanceOf$className.copyWith(...)`${classAnnotation.skipFields ? "" : " or like so:`instanceOf$className.copyWith.fieldName(...)`"}.
       // ignore: library_private_types_in_public_api
-      ${"_\$${classElement.name}CWProxy$typeParametersNames get copyWith => _\$${classElement.name}CWProxyImpl$typeParametersNames(this);"}
+      ${"_\$${className}CWProxy$typeParametersNames get copyWith => _\$${className}CWProxyImpl$typeParametersNames(this);"}
 
       ${classAnnotation.copyWithNull ? _copyWithNullPart(typeAnnotation, sortedFields, classAnnotation.constructor, classAnnotation.skipFields) : ""}
     }
     ''';
   }
 
-  /// Generates the callable class function for copyWith(...).
-  String _copyWithValuesPart(
-    String typeAnnotation,
-    List<FieldInfo> sortedFields,
-    String? constructor,
-    bool skipFields,
-    bool isAbstract,
-  ) {
-    final constructorInput = sortedFields.fold<String>(
-      '',
-      (r, v) {
-        if (v.immutable) return r; // Skip the field
-
-        if (isAbstract) {
-          final type = v.type.endsWith('?') ? v.type : '${v.type}?';
-          return '$r $type ${v.name},';
-        } else {
-          return '$r Object? ${v.name} = const \$CopyWithPlaceholder(),';
-        }
-      },
-    );
-    final paramsInput = sortedFields.fold<String>(
-      '',
-      (r, v) {
-        if (v.immutable) return '$r ${v.name}: _value.${v.name},';
-        final nullCheckForNonNullable =
-            v.nullable ? "" : "|| ${v.name} == null";
-
-        return '''
-        $r ${v.name}:
-        ${v.name} == const \$CopyWithPlaceholder() $nullCheckForNonNullable
-        ? _value.${v.name}
-        // ignore: cast_nullable_to_non_nullable
-        : ${v.name} as ${v.type},''';
-      },
-    );
-
-    final constructorBody = isAbstract
-        ? ""
-        : "{ return ${constructorFor(typeAnnotation, constructor)}($paramsInput); }";
-
-    return '''
-        /// This function **does support** nullification of nullable fields. All `null` values passed to `non-nullable` fields will be ignored.${skipFields ? "" : " You can also use `$typeAnnotation(...).copyWith.fieldName(...)` to override fields one at a time with nullification support."}
-        /// 
-        /// Usage
-        /// ```dart
-        /// $typeAnnotation(...).copyWith(id: 12, name: "My name")
-        /// ````
-        $typeAnnotation call({$constructorInput}) $constructorBody
-    ''';
-  }
-
   /// Generates the complete `copyWithNull` function.
   String _copyWithNullPart(
     String typeAnnotation,
-    List<FieldInfo> sortedFields,
+    List<ConstructorParameterInfo> sortedFields,
     String? constructor,
     bool skipFields,
   ) {
-    /// Return if there is no nullable fields
+    /// Return an empty string when the class has no nullable fields.
     if (sortedFields.where((element) => element.nullable == true).isEmpty) {
       return '';
     }
@@ -117,7 +84,7 @@ class CopyWithGenerator extends GeneratorForAnnotation<CopyWith> {
     final nullConstructorInput = sortedFields.fold<String>(
       '',
       (r, v) {
-        if (v.immutable || !v.nullable) {
+        if (v.fieldAnnotation.immutable || !v.nullable) {
           return r;
         } else {
           return '$r bool ${v.name} = false,';
@@ -127,7 +94,7 @@ class CopyWithGenerator extends GeneratorForAnnotation<CopyWith> {
     final nullParamsInput = sortedFields.fold<String>(
       '',
       (r, v) {
-        if (v.immutable || !v.nullable) {
+        if (v.fieldAnnotation.immutable || !v.nullable) {
           return '$r ${v.name}: ${v.name},';
         } else {
           return '$r ${v.name}: ${v.name} == true ? null : this.${v.name},';
@@ -136,7 +103,7 @@ class CopyWithGenerator extends GeneratorForAnnotation<CopyWith> {
     );
 
     final description = '''
-    /// Copies the object with the specific fields set to `null`. If you pass `false` as a parameter, nothing will be done and it will be ignored. Don't do it. Prefer `copyWith(field: null)`${skipFields ? "" : " or `$typeAnnotation(...).copyWith.fieldName(...)` to override fields one at a time with nullification support"}.
+    /// Copies the object with the specified fields set to `null`. Passing `false` has no effect. Prefer `copyWith(field: null)`${skipFields ? "" : " or `$typeAnnotation(...).copyWith.fieldName(...)` to override fields one at a time with nullification support"}.
     ///
     /// Usage
     /// ```dart
@@ -157,11 +124,12 @@ class CopyWithGenerator extends GeneratorForAnnotation<CopyWith> {
     String type,
     String typeParameters,
     String typeParameterNames,
-    List<FieldInfo> sortedFields,
+    List<ConstructorParameterInfo> sortedFields,
     bool skipFields,
   ) {
     final typeAnnotation = type + typeParameterNames;
-    final filteredFields = sortedFields.where((e) => !e.immutable);
+    final filteredFields =
+        sortedFields.where((e) => !e.fieldAnnotation.immutable);
 
     final nonNullableFunctions = skipFields ? "" : filteredFields.map((e) => '''
     @override
@@ -181,15 +149,66 @@ class CopyWithGenerator extends GeneratorForAnnotation<CopyWith> {
 
       /// Proxy class for `copyWith` functionality. This is a callable class and can be used as follows: `instanceOf$type.copyWith(...)`.${skipFields ? "" : " Additionally contains functions for specific fields e.g. `instanceOf$type.copyWith.fieldName(...)`"}
       class _\$${type}CWProxyImpl$typeParameters implements _\$${type}CWProxy$typeParameterNames {
-        final $type$typeParameterNames _value;
-
         const _\$${type}CWProxyImpl(this._value);
+
+        final $type$typeParameterNames _value;
 
         $nonNullableFunctions
 
         @override
         ${_copyWithValuesPart(typeAnnotation, sortedFields, constructor, skipFields, false)}
       }
+    ''';
+  }
+
+  /// Generates the callable class function for copyWith(...).
+  String _copyWithValuesPart(
+    String typeAnnotation,
+    List<ConstructorParameterInfo> sortedFields,
+    String? constructor,
+    bool skipFields,
+    bool isAbstract,
+  ) {
+    final constructorInput = sortedFields.fold<String>(
+      '',
+      (r, v) {
+        if (v.fieldAnnotation.immutable) return r; // Skip the field
+
+        if (isAbstract) {
+          return '$r ${v.type} ${v.name},';
+        } else {
+          return '$r Object? ${v.name} = const \$CopyWithPlaceholder(),';
+        }
+      },
+    );
+
+    final paramsInput = sortedFields.fold<String>(
+      '',
+      (r, v) {
+        if (v.fieldAnnotation.immutable) {
+          return '$r ${v.name}: _value.${v.name},';
+        }
+
+        return '''$r ${v.isPositioned ? "" : '${v.name}:'}
+        ${v.name} == const \$CopyWithPlaceholder()
+        ? _value.${v.name}
+        // ignore: cast_nullable_to_non_nullable
+        : ${v.name} as ${v.type},''';
+      },
+    );
+
+    final constructorBody = isAbstract
+        ? ""
+        : "{ return ${constructorFor(typeAnnotation, constructor)}($paramsInput); }";
+
+    return '''
+        /// This function **does support** nullification of nullable fields. All `null` values passed to `non-nullable` fields will be ignored.${skipFields ? "" : " You can also use `$typeAnnotation(...).copyWith.fieldName(...)` to override fields one at a time with nullification support."}
+        /// 
+        /// Usage
+        /// ```dart
+        /// $typeAnnotation(...).copyWith(id: 12, name: "My name")
+        /// ````
+        $typeAnnotation call({$constructorInput}) $constructorBody
     ''';
   }
 }