+++ copy_with_extension_gen/lib/src/field_info.dart
@@ -1,27 +1,78 @@
 import 'package:analyzer/dart/constant/value.dart' show DartObject;
 import 'package:analyzer/dart/element/element.dart'
-    show ClassElement, FieldElement, ParameterElement, PrefixElement;
+    show ParameterElement, PrefixElement;
 import 'package:analyzer/dart/element/nullability_suffix.dart';
 import 'package:copy_with_extension/copy_with_extension.dart';
 import 'package:copy_with_extension_gen/src/helpers.dart';
+import 'package:analyzer/dart/element/element2.dart'
+    show ClassElement2, FieldElement2, FormalParameterElement;
+import 'package:copy_with_extension_gen/src/copy_with_field_annotation.dart';
+import 'package:copy_with_extension_gen/src/helpers.dart'
+    show readElementNameOrThrow;
 import 'package:source_gen/source_gen.dart' show ConstantReader, TypeChecker;
 
-/// Represents a single class field with the additional metadata needed for code generation.
+/// Class field info relevant for code generation.
 class FieldInfo {
+  FieldInfo({required this.name, required this.nullable, required this.type});
+
+  /// Parameter / field type.
   final String name;
-  final String type;
-  final bool immutable;
+
+  /// If the type is nullable. `dynamic` is considered non-nullable as it doesn't have nullability flag.
   final bool nullable;
 
-  FieldInfo(ParameterElement element, ClassElement classElement)
-      : name = element.name,
-        type = _fullTypeName(element),
-        immutable = _readFieldAnnotation(element, classElement).immutable,
-        nullable = element.type.nullabilitySuffix != NullabilitySuffix.none;
+  /// Type name with nullability flag.
+  final String type;
+
+  /// True if the type is `dynamic`.
+  bool get isDynamic => type == "dynamic";
+}
+
+/// Represents a single class field with the additional metadata needed for code generation.
+class ConstructorParameterInfo extends FieldInfo {
+  ConstructorParameterInfo(
+    FormalParameterElement element,
+    ClassElement2 classElement, {
+    required this.isPositioned,
+  })  : fieldAnnotation = _readFieldAnnotation(element, classElement),
+        classFieldInfo =
+            _classFieldInfo(readElementNameOrThrow(element), classElement),
+        super(
+          name: readElementNameOrThrow(element),
+          nullable: element.type.nullabilitySuffix != NullabilitySuffix.none,
+          type: element.type.getDisplayString(),
+        );
+
+  /// Annotation provided by the user with `CopyWithField`.
+  final CopyWithFieldAnnotation fieldAnnotation;
+
+  /// True if the field is positioned in the constructor
+  final bool isPositioned;
+
+  /// Info relevant to the given field taken from the class itself, as contrary to the constructor parameter.
+  /// If `null`, the field with the given name wasn't found on the class.
+  final FieldInfo? classFieldInfo;
 
   @override
   String toString() {
-    return 'type:$type name:$name immutable:$immutable nullable:$nullable';
+    return 'type:$type name:$name fieldAnnotation:$fieldAnnotation nullable:$nullable';
+  }
+
+  /// Returns the field info for the constructor parameter in the relevant class.
+  static FieldInfo? _classFieldInfo(
+    String fieldName,
+    ClassElement2 classElement,
+  ) {
+    final field = classElement.fields2
+        .where((e) => readElementNameOrThrow(e) == fieldName)
+        .fold<FieldElement2?>(null, (previousValue, element) => element);
+    if (field == null) return null;
+
+    return FieldInfo(
+      name: readElementNameOrThrow(field),
+      nullable: field.type.nullabilitySuffix != NullabilitySuffix.none,
+      type: field.type.getDisplayString(),
+    );
   }
 
   /// Returns full type name including namespace.
@@ -39,26 +90,29 @@ class FieldInfo {
   }
 
   /// Restores the `CopyWithField` annotation provided by the user.
-  static CopyWithField _readFieldAnnotation(
-    ParameterElement element,
-    ClassElement classElement,
+  static CopyWithFieldAnnotation _readFieldAnnotation(
+    FormalParameterElement element,
+    ClassElement2 classElement,
   ) {
-    final fieldElement = classElement.getField(element.name);
-    if (fieldElement is! FieldElement) {
-      return const CopyWithField();
+    const defaults = CopyWithFieldAnnotation.defaults();
+
+    final fieldElement =
+        classElement.getField2(readElementNameOrThrow(element));
+    if (fieldElement is! FieldElement2) {
+      return defaults;
     }
 
     const checker = TypeChecker.fromRuntime(CopyWithField);
     final annotation = checker.firstAnnotationOf(fieldElement);
     if (annotation is! DartObject) {
-      return const CopyWithField();
+      return defaults;
     }
 
     final reader = ConstantReader(annotation);
-    final immutable = reader.read('immutable').literalValue as bool?;
+    final immutable = reader.peek('immutable')?.boolValue;
 
-    return CopyWithField(
-      immutable: immutable ?? const CopyWithField().immutable,
+    return CopyWithFieldAnnotation(
+      immutable: immutable ?? defaults.immutable,
     );
   }
 }