+++ copy_with_extension_gen/lib/src/helpers.dart
@@ -1,21 +1,22 @@
-import 'package:analyzer/dart/element/element.dart'
-    show ClassElement, ConstructorElement;
-import 'package:copy_with_extension/copy_with_extension.dart';
+import 'package:analyzer/dart/element/element2.dart'
+    show ClassElement2, ConstructorElement2, Element2;
+import 'package:copy_with_extension_gen/src/copy_with_annotation.dart';
 import 'package:copy_with_extension_gen/src/field_info.dart';
+import 'package:copy_with_extension_gen/src/settings.dart';
 import 'package:source_gen/source_gen.dart'
     show ConstantReader, InvalidGenerationSourceError;
 
 /// Generates a list of `FieldInfo` for each class field that will be a part of the code generation process.
 /// The resulting array is sorted by the field name. `Throws` on error.
-List<FieldInfo> sortedConstructorFields(
-  ClassElement element,
+List<ConstructorParameterInfo> sortedConstructorFields(
+  ClassElement2 element,
   String? constructor,
 ) {
   final targetConstructor = constructor != null
-      ? element.getNamedConstructor(constructor)
-      : element.unnamedConstructor;
+      ? element.getNamedConstructor2(constructor)
+      : element.unnamedConstructor2;
 
-  if (targetConstructor is! ConstructorElement) {
+  if (targetConstructor is! ConstructorElement2) {
     if (constructor != null) {
       throw InvalidGenerationSourceError(
         'Named Constructor "$constructor" constructor is missing.',
@@ -23,48 +24,48 @@ List<FieldInfo> sortedConstructorFields(
       );
     } else {
       throw InvalidGenerationSourceError(
-        'Default constructor for "${element.name}" is missing.',
+        'Default constructor for $element is missing.',
         element: element,
       );
     }
   }
 
-  final parameters = targetConstructor.parameters;
+  final parameters = targetConstructor.formalParameters;
   if (parameters.isEmpty) {
     throw InvalidGenerationSourceError(
-      'Unnamed constructor for ${element.name} has no parameters or missing.',
+      'Unnamed constructor for $element has no parameters or missing.',
       element: element,
     );
   }
 
-  for (final parameter in parameters) {
-    if (!parameter.isNamed) {
-      final constructorName = targetConstructor.name.isEmpty
-          ? 'Unnamed constructor'
-          : 'Constructor "${targetConstructor.name}"';
-      throw InvalidGenerationSourceError(
-        '$constructorName for "${element.name}" contains unnamed parameter "${parameter.name}". Constructors annotated with "CopyWith" can contain only named parameters.',
-        element: element,
-      );
-    }
-  }
+  final fields = <ConstructorParameterInfo>[];
 
-  final fields = parameters.map((v) => FieldInfo(v, element)).toList();
-  fields.sort((lhs, rhs) => lhs.name.compareTo(rhs.name));
+  for (final parameter in parameters) {
+    final field = ConstructorParameterInfo(
+      parameter,
+      element,
+      isPositioned: parameter.isPositional,
+    );
+
+    fields.add(field);
+  }
 
   return fields;
 }
 
 /// Restores the `CopyWith` annotation provided by the user.
-CopyWith readClassAnnotation(ConstantReader reader) {
-  final generateCopyWithNull = reader.read('copyWithNull').boolValue;
-  final skipFields = reader.read('skipFields').boolValue;
+CopyWithAnnotation readClassAnnotation(
+  Settings settings,
+  ConstantReader reader,
+) {
+  final generateCopyWithNull = reader.peek('copyWithNull')?.boolValue;
+  final skipFields = reader.peek('skipFields')?.boolValue;
   final constructor = reader.peek('constructor')?.stringValue;
 
-  return CopyWith(
-    copyWithNull: generateCopyWithNull,
+  return CopyWithAnnotation(
+    copyWithNull: generateCopyWithNull ?? settings.copyWithNull,
+    skipFields: skipFields ?? settings.skipFields,
     constructor: constructor,
-    skipFields: skipFields,
   );
 }
 
@@ -73,16 +74,28 @@ CopyWith readClassAnnotation(ConstantReader reader) {
 /// If `nameOnly` is `true`: `class MyClass<T extends String, Y>` returns `<T, Y>`.
 ///
 /// If `nameOnly` is `false`: `class MyClass<T extends String, Y>` returns `<T extends String, Y>`.
-String typeParametersString(ClassElement classElement, bool nameOnly) {
-  final names = classElement.typeParameters
+String typeParametersString(ClassElement2 classElement, bool nameOnly) {
+  final names = classElement.typeParameters2
       .map(
-        (e) => nameOnly ? e.name : e.getDisplayString(withNullability: true),
+        (e) => nameOnly ? readElementNameOrThrow(e) : e.displayString2(),
       )
       .join(',');
   if (names.isNotEmpty) {
     return '<$names>';
   } else {
     return '';
+  }
+}
+
+String readElementNameOrThrow(Element2 element) {
+  final name = element.name3;
+  if (name is String) {
+    return name;
+  } else {
+    throw InvalidGenerationSourceError(
+      'Name for $element is missing.',
+      element: element,
+    );
   }
 }