+++ copy_with_extension_gen/test/generated_code_test_cases/test_cases_exceptions.dart
@@ -3,11 +3,12 @@ part of 'source_gen_entrypoint.dart';
 @ShouldThrow(
   'Only classes can be annotated with "CopyWith". "Object wrongAnnotation" is not a ClassElement.',
 )
+// ignore: invalid_annotation_target
 @CopyWith()
 Object wrongAnnotation = Object();
 
 @ShouldThrow(
-  'Unnamed constructor for NoConstructor has no parameters or missing.',
+  'Unnamed constructor for class NoConstructor has no parameters or missing.',
 )
 @CopyWith()
 class NoConstructor {
@@ -18,41 +19,22 @@ class NoConstructor {
 @CopyWith(constructor: "test")
 class WrongConstructor {}
 
-@ShouldThrow('Default constructor for "NoDefaultConstructor" is missing.')
+@ShouldThrow('Default constructor for class NoDefaultConstructor is missing.')
 @CopyWith()
 class NoDefaultConstructor {
   NoDefaultConstructor.nonDefault();
 }
 
 @ShouldThrow(
-  'Unnamed constructor for "ConstructorWithSomeUnnamedFields" contains unnamed parameter "value". Constructors annotated with "CopyWith" can contain only named parameters.',
+  'The constructor parameter "nullableWithNonNullableConstructor" is not nullable, whereas the corresponding class field is nullable. This use case is not supported.',
 )
 @CopyWith()
-class ConstructorWithSomeUnnamedFields {
-  int? test;
-  int value;
+class TestNullability {
+  TestNullability(
+    int this.nullableWithNonNullableConstructor,
+  );
 
-  ConstructorWithSomeUnnamedFields(this.value, {this.test});
-}
-
-@ShouldThrow(
-  'Unnamed constructor for "ConstructorWithAllUnnamedFields" contains unnamed parameter "value". Constructors annotated with "CopyWith" can contain only named parameters.',
-)
-@CopyWith()
-class ConstructorWithAllUnnamedFields {
-  int? test;
-  int value;
-
-  ConstructorWithAllUnnamedFields(this.value, this.test);
-}
-
-@ShouldThrow(
-  'Constructor "test" for "NamedConstructorWithSomeUnnamedFields" contains unnamed parameter "value". Constructors annotated with "CopyWith" can contain only named parameters.',
-)
-@CopyWith(constructor: "test")
-class NamedConstructorWithSomeUnnamedFields {
-  int? test;
-  int value;
-
-  NamedConstructorWithSomeUnnamedFields.test(this.value, this.test);
+  // Some info on this: https://github.com/numen31337/copy_with_extension/pull/69
+  // If a field is nullable, you can change the type of the constructor parameter to be non-nullable. However, if you do this, an exception may be thrown because the constructor only accepts non-nullable parameters. The issue is that we cannot guarantee that the parameter will be non-null when calling the constructor in the `copyWith` function. Therefore, even if the constructor constructs an object with a nullable field, it is currently impossible to achieve this in the implementation.
+  final int? nullableWithNonNullableConstructor;
 }